Move AppErrorView to the `api` module so that `:appnav:root` does not have to depend on a `impl` module.
