+++ services/apperror/api/build.gradle.kts
@@ -16,4 +16,5 @@ android {
 
 dependencies {
     implementation(libs.coroutines.core)
+    implementation(projects.libraries.designsystem)
 }
